+++ source/daemon/crates/wardnetd/src/system/tests/mod.rs
@@ -1,0 +1,2 @@
+mod pnet_network_probe;
+mod proc_net_inspector;
