+++ source/daemon/crates/wardnetd/src/tests/mod.rs
@@ -2,6 +2,7 @@ mod device_detector;
 mod device_snapshot_listener;
 mod firewall_netlink;
 mod garp_learning;
+mod garp_pnet;
 mod health_runner;
 mod heartbeat;
 mod hostname_resolver;
@@ -16,6 +17,7 @@ mod routing_listener;
 pub mod stubs;
 mod systemctl_power_ops;
 mod tls_server;
+mod tunnel_exit_probe;
 mod tunnel_idle;
 mod tunnel_interface_wireguard;
 mod tunnel_monitor;
