+++ source/daemon/crates/wardnetd/src/tunnel_exit_probe.rs
@@ -79,7 +79,7 @@ impl TunnelExitProbe for ReqwestTunnelExitProbe {
 /// The body is a sequence of `key=value` lines; `ip=` and `loc=` are
 /// the two we need. Missing fields fail with [`ProbeError::Parse`] so
 /// the user sees a precise error rather than a generic "probe failed".
-fn parse_trace_body(body: &str) -> Result<ExitInfo, ProbeError> {
+pub(crate) fn parse_trace_body(body: &str) -> Result<ExitInfo, ProbeError> {
     let mut ip: Option<String> = None;
     let mut country: Option<String> = None;
 
@@ -105,27 +105,3 @@ fn parse_trace_body(body: &str) -> Result<ExitInfo, ProbeError> {
     Ok(ExitInfo { ip, country_code })
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_trace_body_extracts_ip_and_loc() {
-        let body = "fl=53f1\nh=1.1.1.1\nip=203.0.113.42\nts=1714000000.0\nvisit_scheme=https\nuag=test\ncolo=AMS\nsliver=none\nhttp=http/2\nloc=NL\ntls=TLSv1.3\nsni=plaintext\nwarp=off\ngateway=off\nrbi=off\nkex=X25519\n";
-        let info = parse_trace_body(body).expect("parse ok");
-        assert_eq!(info.ip, "203.0.113.42");
-        assert_eq!(info.country_code, "NL");
-    }
-
-    #[test]
-    fn parse_trace_body_missing_ip_errors() {
-        let body = "loc=DE\n";
-        assert!(matches!(parse_trace_body(body), Err(ProbeError::Parse(_))));
-    }
-
-    #[test]
-    fn parse_trace_body_missing_loc_errors() {
-        let body = "ip=203.0.113.1\n";
-        assert!(matches!(parse_trace_body(body), Err(ProbeError::Parse(_))));
-    }
-}
